v1.3.3 Trying to dynamically load content into a loaded html page (app.html) and I'm going nowhere fast. I think its to do with the asynchronosity of the app loading things, but I just can't quite find the answer, though I know I'm very close because my test code shows that the code is working.
