+++ app.html
@@ -1,0 +1,44 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+</head>
+<body>
+
+    <!-- MASTHEAD -->
+    <header class="row justify-content-center">
+        <a href="/index.html" class="col place-content-center navbar">
+            <img class="logo" src="/imgs/mindforge40k-logo-white.svg" alt="Mindforge 40k logo">
+        </a>
+    </header>
+
+    <!-- BACK BUTTON -->
+    <section class="row">
+        <div class="col pb-4">
+            <a class="row gap-2 px-4" href="/index.html" id="go-back-link">
+                <img class="inline-block" src="/imgs/arrow-back.svg" alt="">
+                <p class="dark-grey-2 p-sm inline-block">Go back</p>
+            </a>
+        </div>
+    </section>
+
+    <!-- MAIN -->
+    <section id="main">
+        <p class="no-content white">
+            No content to load
+        </p>
+    </section>
+
+    <!-- FOOTER -->
+    <footer class="justify-content-center bg-dark fixed">
+        <p class="dark-grey-2 p-sm">Coded by <a href="https://benhoath.com" target="_blank" rel="noreferrer noopener">Ben Hoath</a>, developed with <span class="red">♥</span></p>
+    </footer>
+
+    <!-- SCRIPTS -->
+    <script src="/src/head.js"></script>
+    <!-- <script src="/src/gameSetup.js" type="module"></script> -->
+    
+    <script src="/src/app.js" type="module"></script>
+
+
+</body>
+</html>
