+++ index.html
@@ -35,7 +35,7 @@
                 <img src="imgs/icon-free-play-white.svg" type="svg" alt="Free Play icon">
                 <h1>Free Play</h1>
                 <p class="white menu-desc">How many can you correctly guess?</p>
-                <a href="/setup-free-play.html" class="button btn-steel-grey" id="btn-setup-free-play">PLAY</a>
+                <button class="button btn-steel-grey" id="btn-setup-free-play" data-button="setup-free-play">PLAY</button>
             </div>
 
             <!-- TIMED MODE -->
@@ -44,7 +44,7 @@
                 <h1>Timed Mode</h1>
                 <p class="white menu-desc">Guess as many correct answers
                     <br>within the time limit!</p>
-                <a href="/setup-timed-mode.html" class="button btn-steel-grey" id="btn-setup-timed-mode">PLAY</a>
+                <button class="button btn-steel-grey" id="btn-setup-timed-mode" data-load="app" data-button="setup-timed-mode">PLAY</button>
             </div>
 
             <!-- SUDDEN DEATH -->
@@ -52,7 +52,7 @@
                 <img src="imgs/icon-sudden-death-white.svg" type="svg" alt="Sudden Death icon">
                 <h1>Sudden Death</h1>
                 <p class="white menu-desc">Guess incorrectly and end the game. <br>How many can you score?</p>
-                <a href="#" class="button btn-disabled" id="btn-setup-sudden-death">PLAY</a>
+                <button class="button btn-disabled" id="btn-setup-sudden-death">PLAY</button>
             </div>
 
         </div>
@@ -66,7 +66,7 @@
     </section>
 
     <!-- High score -->
-    <section class="justify-content-center overflow">
+    <!-- <section class="justify-content-center overflow">
         <div class="row slider-wrapper" id="highscores-wrapper">
 
             <div class="col px-8 py-8 gap-4 full-width highscores-slide" id="highscores-free-play">
@@ -136,11 +136,11 @@
                 <span class="pagination-dot" id="highscore-pag3"></span>
             </div>
         </div>
-    </section>
+    </section> -->
 
     <footer class="justify-content-center bg-dark fixed">
         <p class="dark-grey-2 p-sm">Coded by <a href="https://benhoath.com" target="_blank" rel="noreferrer noopener">Ben Hoath</a>, developed with <span class="red">♥</span></p>
     </footer>
-    <script src="/src/app.js"></script>
+    <script src="/src/index.js" type="module"></script>
 </body>
 </html>
