+++ styles.css
@@ -571,3 +571,8 @@ footer > p > a:visited {
 .strong {
     font-weight: bold;
 }
+
+.no-content {
+    display: flex;
+    justify-content: center;
+}
